requirements.txt
fastapi
uvicorn
httpx
pyserial
sqlalchemy
pytest
pydantic-settings
pyyaml
pytest-asyncio
email-validator
